requirements.txt
flask==2.0.1
flask-cors==3.0.10
scikit-learn==0.24.2
pandas==1.3.3
numpy==1.21.2
joblib==1.0.2 
Pillow==8.3.2
# Removed scikit-image and OpenCV to ensure compatibility with Python 3.13 and simplify dependencies